requirements.txt
altgraph==0.17.4
packaging==25.0
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
setuptools==80.9.0
